+++ kopy-compiler/main/kotlin/com/javiersc/kotlin/kopy/compiler/ir/transformers/IrAtomicPropertyTransformer.kt
@@ -3,6 +3,7 @@ package com.javiersc.kotlin.kopy.compiler.ir.transformers
 import com.javiersc.kotlin.compiler.extensions.ir.declarationIrBuilder
 import com.javiersc.kotlin.kopy.compiler.atomicReferenceClassId
 import com.javiersc.kotlin.kopy.compiler.underscoreAtomicName
+import org.jetbrains.kotlin.DeprecatedForRemovalCompilerApi
 import org.jetbrains.kotlin.backend.common.IrElementTransformerVoidWithContext
 import org.jetbrains.kotlin.backend.common.extensions.IrPluginContext
 import org.jetbrains.kotlin.ir.IrStatement
@@ -28,6 +29,8 @@ internal class IrAtomicPropertyTransformer(
     private val pluginContext: IrPluginContext,
 ) : IrElementTransformerVoidWithContext() {
 
+    // TODO: Remove @OptIn(DeprecatedForRemovalCompilerApi::class)
+    @OptIn(DeprecatedForRemovalCompilerApi::class)
     override fun visitPropertyNew(declaration: IrProperty): IrStatement {
         fun originalProp(): IrStatement = super.visitPropertyNew(declaration)
         if (declaration.name != underscoreAtomicName) return originalProp()