+++ kopy-compiler/test-gen/java/com/javiersc/kotlin/kopy/compiler/Kopy1PublicDiagnosticTestGenerated.java
@@ -4,6 +4,7 @@ package com.javiersc.kotlin.kopy.compiler;
 
 import com.intellij.testFramework.TestDataPath;
 import org.jetbrains.kotlin.test.util.KtTestUtil;
+import org.jetbrains.kotlin.test.TargetBackend;
 import org.jetbrains.kotlin.test.TestMetadata;
 import org.junit.jupiter.api.Test;
 
@@ -17,7 +18,7 @@ import java.util.regex.Pattern;
 public class Kopy1PublicDiagnosticTestGenerated extends AbstractKopy1PublicDiagnosticTest {
   @Test
   public void testAllFilesPresentIn1_public() {
-    KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("test-data/diagnostics-kopy-visibility/1_public"), Pattern.compile("^(.+)\\.kt$"), null, true);
+    KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("test-data/diagnostics-kopy-visibility/1_public"), Pattern.compile("^(.+)\\.kt$"), null, TargetBackend.JVM_IR, true);
   }
 
   @Test