+++ kopy-compiler/test-gen/java/com/javiersc/kotlin/kopy/compiler/Kopy3ProtectedDiagnosticTestGenerated.java
@@ -4,6 +4,7 @@ package com.javiersc.kotlin.kopy.compiler;
 
 import com.intellij.testFramework.TestDataPath;
 import org.jetbrains.kotlin.test.util.KtTestUtil;
+import org.jetbrains.kotlin.test.TargetBackend;
 import org.jetbrains.kotlin.test.TestMetadata;
 import org.junit.jupiter.api.Test;
 
@@ -17,7 +18,7 @@ import java.util.regex.Pattern;
 public class Kopy3ProtectedDiagnosticTestGenerated extends AbstractKopy3ProtectedDiagnosticTest {
   @Test
   public void testAllFilesPresentIn3_protected() {
-    KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("test-data/diagnostics-kopy-visibility/3_protected"), Pattern.compile("^(.+)\\.kt$"), null, true);
+    KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("test-data/diagnostics-kopy-visibility/3_protected"), Pattern.compile("^(.+)\\.kt$"), null, TargetBackend.JVM_IR, true);
   }
 
   @Test